Document test packages
Bun has a test suite built in, but I noticed we have jest dependencies.
It looks like this is because were are using some jest tooling for the
browser E2E tests.  Rather than refactor this I'd like to document it
for now.  Maybe later it will be worth changing.

Issue #13 Clean package.json